feat(file_renamer.py add DIR check)

+++ file_renamer.py
@@ -1,17 +1,20 @@
 import sys, os
+from pathlib import Path
 
 def renamer (path):
-        path = sys.argv[1]
         dir = os.chdir(path)
         setFilename = "TextFile"
-
         i = 0
-        for filename in os.listdir(dir):         
-                src = path + filename
-                dst = setFilename + str(i) + ".txt"
-                dst = path + dst 
-                
-                os.rename(src, dst) 
-                i += 1
-  
+        
+        if os.path.exists(path):
+                for filename in os.listdir(dir):         
+                        src = path + filename
+                        dst = setFilename + str(i) + ".txt"
+                        dst = path + dst 
+              
+                        os.rename(src, dst) 
+                        i += 1
+        else:
+                print("Directory does not exist!")
+        
 renamer(sys.argv[1])
